backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Google Drive API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Meilisearch
meilisearch==0.31.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
Pillow==10.1.0

mysql-connector-python==8.2.0

# Text Extraction
pytesseract==0.3.10
pdf2image==1.16.3

# Utilities
requests==2.31.0
charset-normalizer==3.3.2

pymssql==2.2.11